_smiles2pubchem to shorten molconvert
