+++ tdc/chemutils/__init__.py
@@ -2,3 +2,4 @@
 from .evaluator import *
 from .molconvert import * 
 from .oracle import * 
+from ._smiles2pubchem import smiles2pubchem
